Optimize string finding and filtering, update README.md, requirements.txt, and add string decryption feature.
