--- sharingan/core/StrFinder/__init__.py
@@ -1,1 +1,0 @@
-__version__ = '2.1.0-detection-only'
